requirements.txt
streamlit
pillow
numpy
pandas
scikit-learn
tensorflow==2.18.0
gdown


